# backend/api/gen/go/admin/shop_banner_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/shop_banner.proto

package admin

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	wrapperspb "google.golang.org/protobuf/types/known/wrapperspb"
	common "shop/api/gen/go/common"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationShopBannerServiceCreateShopBanner = "/admin.ShopBannerService/CreateShopBanner"
const OperationShopBannerServiceDeleteShopBanner = "/admin.ShopBannerService/DeleteShopBanner"
const OperationShopBannerServiceGetShopBanner = "/admin.ShopBannerService/GetShopBanner"
const OperationShopBannerServicePageShopBanner = "/admin.ShopBannerService/PageShopBanner"
const OperationShopBannerServiceSetShopBannerStatus = "/admin.ShopBannerService/SetShopBannerStatus"
const OperationShopBannerServiceUpdateShopBanner = "/admin.ShopBannerService/UpdateShopBanner"

type ShopBannerServiceHTTPServer interface {
	// CreateShopBanner 创建商城轮播图
	CreateShopBanner(context.Context, *ShopBannerForm) (*emptypb.Empty, error)
	// DeleteShopBanner 删除商城轮播图
	DeleteShopBanner(context.Context, *wrapperspb.StringValue) (*emptypb.Empty, error)
	// GetShopBanner 查询商城轮播图
	GetShopBanner(context.Context, *wrapperspb.Int64Value) (*ShopBannerForm, error)
	// PageShopBanner 查询商城轮播图分页列表
	PageShopBanner(context.Context, *PageShopBannerRequest) (*PageShopBannerResponse, error)
	// SetShopBannerStatus 设置状态
	SetShopBannerStatus(context.Context, *common.SetStatusRequest) (*emptypb.Empty, error)
	// UpdateShopBanner 更新商城轮播图
	UpdateShopBanner(context.Context, *ShopBannerForm) (*emptypb.Empty, error)
}

func RegisterShopBannerServiceHTTPServer(s *http.Server, srv ShopBannerServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/admin/shop/banner", _ShopBannerService_PageShopBanner0_HTTP_Handler(srv))
	r.GET("/api/admin/shop/banner/{value}", _ShopBannerService_GetShopBanner0_HTTP_Handler(srv))
	r.POST("/api/admin/shop/banner", _ShopBannerService_CreateShopBanner0_HTTP_Handler(srv))
	r.PUT("/api/admin/shop/banner/{id}", _ShopBannerService_UpdateShopBanner0_HTTP_Handler(srv))
	r.DELETE("/api/admin/shop/banner/{value}", _ShopBannerService_DeleteShopBanner0_HTTP_Handler(srv))
	r.PUT("/api/admin/shop/banner/{id}/status", _ShopBannerService_SetShopBannerStatus0_HTTP_Handler(srv))
}

func _ShopBannerService_PageShopBanner0_HTTP_Handler(srv ShopBannerServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PageShopBannerRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopBannerServicePageShopBanner)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PageShopBanner(ctx, req.(*PageShopBannerRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PageShopBannerResponse)
		return ctx.Result(200, reply)
	}
}

func _ShopBannerService_GetShopBanner0_HTTP_Handler(srv ShopBannerServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in wrapperspb.Int64Value
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopBannerServiceGetShopBanner)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetShopBanner(ctx, req.(*wrapperspb.Int64Value))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ShopBannerForm)
		return ctx.Result(200, reply)
	}
}

func _ShopBannerService_CreateShopBanner0_HTTP_Handler(srv ShopBannerServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ShopBannerForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopBannerServiceCreateShopBanner)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateShopBanner(ctx, req.(*ShopBannerForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ShopBannerService_UpdateShopBanner0_HTTP_Handler(srv ShopBannerServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ShopBannerForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopBannerServiceUpdateShopBanner)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateShopBanner(ctx, req.(*ShopBannerForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ShopBannerService_DeleteShopBanner0_HTTP_Handler(srv ShopBannerServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in wrapperspb.StringValue
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopBannerServiceDeleteShopBanner)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteShopBanner(ctx, req.(*wrapperspb.StringValue))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ShopBannerService_SetShopBannerStatus0_HTTP_Handler(srv ShopBannerServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in common.SetStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationShopBannerServiceSetShopBannerStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SetShopBannerStatus(ctx, req.(*common.SetStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type ShopBannerServiceHTTPClient interface {
	// CreateShopBanner 创建商城轮播图
	CreateShopBanner(ctx context.Context, req *ShopBannerForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// DeleteShopBanner 删除商城轮播图
	DeleteShopBanner(ctx context.Context, req *wrapperspb.StringValue, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// GetShopBanner 查询商城轮播图
	GetShopBanner(ctx context.Context, req *wrapperspb.Int64Value, opts ...http.CallOption) (rsp *ShopBannerForm, err error)
	// PageShopBanner 查询商城轮播图分页列表
	PageShopBanner(ctx context.Context, req *PageShopBannerRequest, opts ...http.CallOption) (rsp *PageShopBannerResponse, err error)
	// SetShopBannerStatus 设置状态
	SetShopBannerStatus(ctx context.Context, req *common.SetStatusRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// UpdateShopBanner 更新商城轮播图
	UpdateShopBanner(ctx context.Context, req *ShopBannerForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type ShopBannerServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewShopBannerServiceHTTPClient(client *http.Client) ShopBannerServiceHTTPClient {
	return &ShopBannerServiceHTTPClientImpl{client}
}

// CreateShopBanner 创建商城轮播图
func (c *ShopBannerServiceHTTPClientImpl) CreateShopBanner(ctx context.Context, in *ShopBannerForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/shop/banner"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationShopBannerServiceCreateShopBanner))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteShopBanner 删除商城轮播图
func (c *ShopBannerServiceHTTPClientImpl) DeleteShopBanner(ctx context.Context, in *wrapperspb.StringValue, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/shop/banner/{value}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationShopBannerServiceDeleteShopBanner))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetShopBanner 查询商城轮播图
func (c *ShopBannerServiceHTTPClientImpl) GetShopBanner(ctx context.Context, in *wrapperspb.Int64Value, opts ...http.CallOption) (*ShopBannerForm, error) {
	var out ShopBannerForm
	pattern := "/api/admin/shop/banner/{value}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationShopBannerServiceGetShopBanner))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PageShopBanner 查询商城轮播图分页列表
func (c *ShopBannerServiceHTTPClientImpl) PageShopBanner(ctx context.Context, in *PageShopBannerRequest, opts ...http.CallOption) (*PageShopBannerResponse, error) {
	var out PageShopBannerResponse
	pattern := "/api/admin/shop/banner"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationShopBannerServicePageShopBanner))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SetShopBannerStatus 设置状态
func (c *ShopBannerServiceHTTPClientImpl) SetShopBannerStatus(ctx context.Context, in *common.SetStatusRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/shop/banner/{id}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationShopBannerServiceSetShopBannerStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateShopBanner 更新商城轮播图
func (c *ShopBannerServiceHTTPClientImpl) UpdateShopBanner(ctx context.Context, in *ShopBannerForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/shop/banner/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationShopBannerServiceUpdateShopBanner))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}