# backend/internal/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/admin/shop/banner:
        get:
            tags:
                - ShopBannerService
            description: 查询商城轮播图分页列表
            operationId: ShopBannerService_PageShopBanner
            parameters:
                - name: site
                  in: query
                  description: 位置：枚举【ShopBannerSite】
                  schema:
                    enum:
                        - UNKNOWN_SBS
                        - INDEX
                        - CATEGORY
                    type: string
                    format: enum
                - name: type
                  in: query
                  description: 跳转类型：枚举【ShopBannerType】
                  schema:
                    enum:
                        - UNKNOWN_SBT
                        - GOODS_DETAIL
                        - CATEGORY_DETAIL
                        - WEB_VIEW
                        - MINI
                    type: string
                    format: enum
                - name: status
                  in: query
                  description: 状态
                  schema:
                    enum:
                        - UNKNOWN
                        - ENABLE
                        - DISABLE
                    type: string
                    format: enum
                - name: pageNum
                  in: query
                  description: 当前页码
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  description: 每一页的行数
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PageShopBannerResponse'
        post:
            tags:
                - ShopBannerService
            description: 创建商城轮播图
            operationId: ShopBannerService_CreateShopBanner
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ShopBannerForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/shop/banner/{id}:
        put:
            tags:
                - ShopBannerService
            description: 更新商城轮播图
            operationId: ShopBannerService_UpdateShopBanner
            parameters:
                - name: id
                  in: path
                  description: 主键id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ShopBannerForm'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/shop/banner/{id}/status:
        put:
            tags:
                - ShopBannerService
            description: 设置状态
            operationId: ShopBannerService_SetShopBannerStatus
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SetStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /api/admin/shop/banner/{value}:
        get:
            tags:
                - ShopBannerService
            description: 查询商城轮播图
            operationId: ShopBannerService_GetShopBanner
            parameters:
                - name: value
                  in: path
                  description: The int64 value.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ShopBannerForm'
        delete:
            tags:
                - ShopBannerService
            description: 删除商城轮播图
            operationId: ShopBannerService_DeleteShopBanner
            parameters:
                - name: value
                  in: path
                  description: The string value.
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
...
    /api/app/shop/banner:
        get:
            tags:
                - ShopBannerService
            description: 查询商城轮播图列表
            operationId: ShopBannerService_ListShopBanner
            parameters:
                - name: site
                  in: query
                  description: 位置：枚举【ShopBannerSite】
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListShopBannerResponse'
...
components:
    schemas:
...
        ListShopBannerResponse:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/ShopBanner'
                    description: 分页数据
...
        PageShopBannerResponse:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/ShopBanner'
                    description: 分页数据
                total:
                    type: integer
                    description: 总数
                    format: int32
...
        SetStatusRequest:
            type: object
            properties:
                id:
                    type: string
                    description: 主键id
                status:
                    type: integer
                    description: 状态
                    format: int32
...
        ShopBanner:
            type: object
            properties:
                id:
                    type: string
                    description: 主键id
                site:
                    enum:
                        - UNKNOWN_SBS
                        - INDEX
                        - CATEGORY
                    type: string
                    description: 位置：枚举【ShopBannerSite】
                    format: enum
                picture:
                    type: string
                    description: 图片链接
                type:
                    enum:
                        - UNKNOWN_SBT
                        - GOODS_DETAIL
                        - CATEGORY_DETAIL
                        - WEB_VIEW
                        - MINI
                    type: string
                    description: 跳转类型：枚举【ShopBannerType】
                    format: enum
                href:
                    type: string
                    description: 跳转地址
                sort:
                    type: integer
                    description: 排序
                    format: int32
                status:
                    enum:
                        - UNKNOWN
                        - ENABLE
                        - DISABLE
                    type: string
                    description: 状态
                    format: enum
                createdAt:
                    type: string
                    description: 创建时间
                updatedAt:
                    type: string
                    description: 更新时间
        ShopBannerForm:
            type: object
            properties:
                id:
                    type: string
                    description: 主键id
                site:
                    enum:
                        - UNKNOWN_SBS
                        - INDEX
                        - CATEGORY
                    type: string
                    description: 位置：枚举【ShopBannerSite】
                    format: enum
                picture:
                    type: string
                    description: 图片链接
                type:
                    enum:
                        - UNKNOWN_SBT
                        - GOODS_DETAIL
                        - CATEGORY_DETAIL
                        - WEB_VIEW
                        - MINI
                    type: string
                    description: 跳转类型：枚举【ShopBannerType】
                    format: enum
                href:
                    type: string
                    description: 跳转地址
                sort:
                    type: integer
                    description: 排序
                    format: int32
                status:
                    enum:
                        - UNKNOWN
                        - ENABLE
                        - DISABLE
                    type: string
                    description: 状态：枚举【Status】
                    format: enum
...